api/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
sqlalchemy==2.0.23
alembic==1.13.0
chromadb==0.4.18
httpx==0.25.2
python-dotenv==1.0.0
textstat==0.7.3
spacy==3.7.2
nltk==3.8.1
sentence-transformers==2.2.2
rapidfuzz==3.5.2
GitPython==3.1.40
typer==0.9.0
python-frontmatter==1.1.0